Did some more gobbledigook with Connect Four

+++ cogs/c4.py
@@ -4,6 +4,7 @@ Written for Python 3.6 and discord.py 1.0.0a"""
 import discord
 from discord.ext import commands
 from .utils import checks
+from main import app_name
 
 
 class Session:
@@ -54,7 +55,7 @@ class Session:
         for row in self.board:
             board.append([self.emojis[i] for i in row[:-1]][::-1])
 
-        return board
+        return "\n".join(["{}{}{}{}{}{}{}".format(*[board[y][x] for y in range(7)]) for x in range(6)])
 
     def check(self, board):
         for x in range(7):
@@ -89,48 +90,54 @@ class Session:
                                 board[x][y] == board[x - 3][y + 3]:
                     return self.p1 if board[x][y] == self.p1.id else self.p2
 
-        if all(map(lambda x: x == 6, [r[6] for r in self.board])):
+        if all(map(lambda n: n == 6, [r[6] for r in self.board])):
             return "Draw"
 
         return False
 
 
 class ConnectFour:
-    """Play a game of Connect Four"""
+    """Play a game of Connect Four
+
+    See the help manual for individual
+    commands for more information."""
     def __init__(self, bot):
         self.bot = bot
+        self.db = bot.db
         self.sessions = {}
+        self.message_icon = ["ℹ", "⚠", "⛔"]  # :information_source:, :warning:, :no_entry:
+        self.message_color = [discord.Colour.blue(), discord.Colour.orange(), discord.Colour.red()]
 
     def session(self, ctx):
         return self.sessions.get(ctx.channel.id, None)
 
-    async def c4_error(self, ctx, msg="Error"):
-        em = discord.Embed(description=f"⛔ {msg}", color=discord.Colour.red())
+    def chan_check(self, ctx):
+        return str(ctx.channel.id) in self.db.smembers(f"{app_name}:c4:allowed_channels")
+
+    async def message(self, ctx, msg="Error", level=0):
+        em = discord.Embed(description=f"{self.message_icon[level]}  {msg}", color=self.message_color[level])
         await ctx.send(embed=em)
 
     async def send_board(self, ctx, init=False, win=False):
         session = self.session(ctx)
         if session.msg is not None:
-            try:
-                await session.msg.delete()
-            except Exception as e:
-                await ctx.send(f"{type(e)}: {e}")
-
-        board = session.get_board
-        parsed_board = "\n".join(["{}{}{}{}{}{}{}".format(*[board[y][x] for y in range(7)]) for x in range(6)])
+            await session.msg.delete()
 
         if win:
             if win == "Draw":
                 turn = f"Game ended in a Draw."
+                color = discord.Colour.dark_grey()
             else:
                 turn = f"Game Over!\n{win.name} wins! 🎉"
+                color = 0xFDFF00
         else:
             turn = "New game!" if init else f"Turn: {(session.turn + 2) // 2}"
+            color = session.color
 
         em = discord.Embed(title=f"{session.player_chip(session.p1)}{session.p1.name} 🆚 "
                                  f"{session.p2.name}{session.player_chip(session.p2)}",
-                           description=f"{turn}\n\n:one::two::three::four::five::six::seven:\n{parsed_board}",
-                           color=session.color)
+                           description=f"{turn}\n\n:one::two::three::four::five::six::seven:\n{session.get_board}",
+                           color=color)
 
         if win:
             self.sessions.pop(ctx.channel.id)
@@ -139,37 +146,67 @@ class ConnectFour:
             em.set_footer(text=f"{session.current_player.name}'s turn: {session.current_player_chip}")
             session.msg = await ctx.send(embed=em)
 
-        await ctx.message.delete()
+        if ctx.channel.permissions_for(ctx.guild.me).manage_messages:
+            await ctx.message.delete()
 
-    @commands.group()
-    async def c4(self, ctx):
+    @commands.group(name="play")
+    async def play(self, ctx):
         """Connect Four
 
         The classic game of Connect Four.
         Use these commands to play a game
         of Connect Four with another user.
         You can have multiple concurrent
-        games, one per channel."""
+        games, one per channel.
+
+        Use `[p]c4 <column>` to place a piece.
+        First to get four in a row, whether
+        horizontal, vertical, or diagonal wins."""
+        if not self.chan_check(ctx):
+            return
         if not ctx.invoked_subcommand:
             await self.bot.formatter.format_help_for(ctx, ctx.command)
 
-    @c4.command(name="start", aliases=["play"])
-    async def _start(self, ctx, *, user: discord.Member=None):
+    @play.command(name="c4")
+    async def _c4(self, ctx, *, user: discord.Member=None):
         """Star a game of Connect Four
 
         `[p]c4 start @user` will start a game
         with that user in the current channel."""
-        # if user:
-        #     await ctx.send(f"Ping! Confirmed user: {user.name} (Currently not implemented)")
+        if not self.chan_check(ctx):
+            return
         if user:
             session = self.session(ctx)
             if session:
-                await self.c4_error(ctx, msg="There is already an active game in this channel.")
+                await self.message(ctx, msg="There is already an active game in this channel.", level=2)
             else:
                 self.sessions[ctx.channel.id] = Session(ctx.author, user, ctx.channel)
                 await self.send_board(ctx)
         else:
             await self.bot.formatter.format_help_for(ctx, ctx.command, "You need another player to start.")
+
+    @commands.group(name="c4", invoke_without_command=True)
+    async def c4(self, ctx, *, arg: int=None):
+        """Make a Move
+
+        `[p]c4 <column>` will place a chip
+        in that row. Only the current player
+        can use this command."""
+        if not self.chan_check(ctx):
+            return
+        ctx.message.content = f"{self.bot.command_prefix[0]}c4 move {arg}"
+        await self.bot.process_commands(ctx.message)
+
+    @c4.command(name="play", hidden=True)
+    async def _play(self, ctx, *, user: discord.Member=None):
+        """Star a game of Connect Four
+
+        `[p]c4 start @user` will start a game
+        with that user in the current channel."""
+        if not self.chan_check(ctx):
+            return
+        ctx.message.content = f"{self.bot.command_prefix[0]}play c4 {user.mention}"
+        await self.bot.process_commands(ctx.message)
 
     @c4.command(name="quit", aliases=["end"])
     async def _quit(self, ctx):
@@ -177,50 +214,37 @@ class ConnectFour:
 
         `[p]c4 quit` can be used by either player
         to quit their game in the channel."""
-        session = self.sessions.get(ctx.channel.id, None)
+        if not self.chan_check(ctx):
+            return
+        session = self.session(ctx)
         if session and ctx.author in [session.p1, session.p2]:
             self.sessions.pop(ctx.channel.id)
             await ctx.send("Game has ended.")
         else:
-            await self.c4_error(ctx, msg="No active game in this channel.")
+            await self.message(ctx, msg="No active game in this channel.", level=1)
 
     @c4.command(name="board")
     async def board(self, ctx):
         """Resends the current board
 
         Can be used if the board gets lost in the chat"""
+        if not self.chan_check(ctx):
+            return
         session = self.sessions.get(ctx.channel.id, None)
         if session and ctx.author in [session.p1, session.p2]:
             await self.send_board(ctx)
         else:
-            await self.c4_error(ctx, msg="No active game in this channel.")
-
-    @checks.sudo()
-    @c4.command(name="kill", aliases=["killall"])
-    async def _kill(self, ctx):
-        """Aministrative kill command
-
-        This will kill all running games of
-        Connect Four in all channels."""
-        em = discord.Embed(description=f"All running games have been terminated. (Total: {len(self.sessions.keys())})",
-                           color=discord.Colour.dark_grey())
-        self.sessions = {}
-        await ctx.send(embed=em)
-
-    @checks.sudo()
-    @c4.command(name="games")
-    async def games(self, ctx):
-        em = discord.Embed(description=f"Total running games: {len(self.sessions.keys())}",
-                           color=discord.Colour.dark_grey())
-        await ctx.send(embed=em)
+            await self.message(ctx, msg="No active game in this channel.", level=1)
 
     @c4.command(name="move")
     async def _move(self, ctx, row: int):
         """Make a Move
 
-        `[p]c4 move <row>` will place a chip
+        `[p]c4 <column>` will place a chip
         in that row. Only the current player
         can use this command."""
+        if not self.chan_check(ctx):
+            return
         row -= 1
         session = self.session(ctx)
         if session:
@@ -230,11 +254,58 @@ class ConnectFour:
                         check = session.play(ctx.author.id, row)
                         await self.send_board(ctx, win=check)
                     else:
-                        await self.c4_error(ctx, "That row is full. Select another.")
+                        await self.message(ctx, "That row is full. Select another.", level=2)
                 else:
-                    await self.c4_error(ctx, "Invalid row number. Select another.")
+                    await self.message(ctx, "Invalid row number. Select another.", level=2)
         else:
-            await self.c4_error(ctx, msg="No active game in this channel.")
+            await self.message(ctx, msg="No active game in this channel.", level=1)
+
+    @checks.sudo()
+    @c4.command(name="enable", hidden=True)
+    async def _enable(self, ctx, *, chan: discord.TextChannel=None):
+        if not chan:
+            chan = ctx.channel
+        if self.db.sadd(f"{app_name}:c4:allowed_channels", chan.id):
+            await self.message(ctx, msg="Connect Four successfully enabled on channel.")
+        else:
+            await self.message(ctx, msg="Connect Four already enabled on channel.", level=1)
+
+    @checks.sudo()
+    @c4.command(name="disable", hidden=True)
+    async def _disable(self, ctx, *, chan: discord.TextChannel=None):
+        if not chan:
+            chan = ctx.channel
+        if self.db.srem(f"{app_name}:c4:allowed_channels", chan.id):
+            await self.message(ctx, msg="Connect Four successfully disabled on channel.")
+        else:
+            await self.message(ctx, msg="Connect Four already disabled on channel.", level=1)
+
+    @checks.sudo()
+    @c4.command(name="games", hidden=True)
+    async def games(self, ctx):
+        await self.message(ctx, msg=f"Total running games: {len(self.sessions.keys())}")
+
+    @checks.sudo()
+    @c4.command(name="kill", hidden=True)
+    async def _kill(self, ctx, *, _all: bool=False):
+        """Aministrative kill command
+
+        This will kill all running games of
+        Connect Four in all channels."""
+        if not _all:
+            if self.sessions.pop(ctx.channel.id, None):
+                await self.message(ctx, msg="Current game in this channel has been terminated.")
+            else:
+                await self.message(ctx, msg="No active game in this channel.", level=1)
+        else:
+            await self.message(ctx, msg=f"All running games have been terminated. (Total: {len(self.sessions.keys())})")
+            self.sessions = {}
+
+    @checks.sudo()
+    @c4.command(name="killall", hidden=True)
+    async def _killall(self, ctx):
+        ctx.message.content = f"{self.bot.command_prefix[0]}kill True"
+        await self.bot.process_commands(ctx.message)
 
 
 def setup(bot):
